requirements.txt
streamlit
torch
torchvision
timm
numpy
opencv-python
matplotlib
scikit-learn
pyyaml
Pillow
